acadex-backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
alembic>=1.13.1
asyncpg>=0.29.0
pydantic[email]>=2.7.0
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-genai>=0.2.0
httpx>=0.27.0
python-multipart>=0.0.9
google-auth>=2.29.0
Pillow>=10.3.0
